Does not compile.
Either I'm trying to use the logger wrong or it needs an overloaded operator + copy construtor.

Since we CANNOT create a shitload of instances of the logger, we MUST pass a reference to one instance of it throughout our code.  This is true because 2 threads CANNOT write to the same file at the same time and we need one instance to keep it open or we must open and close the damn thing with locking which would suck.