Add fbClient implementation.
fbClient will handle "parsing" the http request, determining what the user wants.

We open a FILE pointer so we can use printf instead of write system calls to make this bearable.

There will be another class that handles the server end  fbHttpServer ?  That class will work with each client request in a multithreaded fashion.. that is the point we break into threads.

Ideally we'll start a copy of fbHttpServer from the main function after forking.  It will setup and then start a background thread that will handle fetching the client requests and executing them on seperate threads.   This will allow the main process to do the backup tasks, etc.