This needs more work.
I'm now testing for a ?.  if there is a ? we know it's a get request for parameters.

The next step is to break up those params into an argument array.

Then we test the url to find out what action it's taking.

Currently, i'm using 404s until handlers are written.

I need fededback on what the scheduler and config stuff needs changed and code to print the current tasks