Ok, so finally figured out the bug with libarchive on linux... it assumes 64 bit offsets for struct stat, instead of default 32 bit. This is fixed in the Makefile. (see http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=419793)
Got rid of test archives in fbCore. We can now use the web interface for
to test with. Modified fbBackup slightly.

Luke, I'm not comfortable just throwing in Chris Reinhart's code when our
code is fine (with the new fix). The ONLY thing he adds as for functionality
is reading symlinks, which are pointless to back up, and the code to do so
probably isn't even copyrightable, since it's standard libarchive interface
and syscalls, making it pointless to include his copyright header. Plus,
since I have no particular copyright notice or license for my code, it
makes it look as though Chris did all the work, and I did nothing, even though
his code adds no significant functionality or fix to the existing code...
Some of the license stuff is getting out of control... Anyway, that's my thought on the matter.